Executable version of code

+++ src/index.ts
@@ -1,17 +1,26 @@
-import {downloadResource, getResource} from "./api/spigot";
-
 require('dotenv').config()
+require('colors')
+
+import type from 'colors'
 
 import yaml from 'yaml'
 import crypto from 'crypto'
 import fs from 'fs'
+import os from 'os'
 import { downloadBuildJar, getBuilds } from "./api/papermc"
 import { spawn } from 'child_process'
 import {downloadAssetJar, getReleases} from "./api/essentialsx";
+import {adoptArch, adoptOS, downloadJRE, getAssets} from "./api/adoptium";
 import {downloadFloodgate, downloadGeyser, getLatestFloodgate, getLatestGeyser} from "./api/geyser";
+import {downloadResource, getResource} from "./api/spigot";
 import {waitForKey} from "./utils/input"
 
+// TODO: get this into an import, without the IDE giving a fucking error
+const decompress = require('decompress')
+const { textSync } = require('figlet')
+
 type serverConfig = {
+  minecraft_version: string,
   plugins: {
     essentials_x: string[],
     other: {
@@ -24,12 +33,33 @@ type serverConfig = {
   }
 }
 
+if (!fs.existsSync('./server-config.yml')) {
+  // Create a default server config file
+  fs.writeFileSync('./server-config.yml', yaml.stringify({
+    minecraft_version: '1.20.1',
+    plugins: {
+      essentials_x: [
+        'EssentialsX',
+        'EssentialsXChat',
+        'EssentialsXGeoIP',
+        'EssentialsXSpawn'
+      ],
+      other: {
+        g_sit: {
+          name: 'GSit',
+          source: 'spigot',
+          id: 62325
+        }
+      }
+    }
+  }))
+}
 
 const config: serverConfig = yaml.parse(fs.readFileSync('./server-config.yml').toString())
-const minecraftVersion: string = '1.20.1' // TODO: Move this to somewhere that isn't code.
 const serverFilesPath: string = './server'
 
 if (!fs.existsSync(serverFilesPath)) fs.mkdirSync(serverFilesPath)
+if (!fs.existsSync('./bin')) fs.mkdirSync('./bin')
 if (!fs.existsSync('./bin/papermc')) fs.mkdirSync('./bin/papermc')
 if (!fs.existsSync('./bin/plugins')) fs.mkdirSync('./bin/plugins')
 
@@ -118,7 +148,7 @@ async function checkForOtherPlugins (): Promise<boolean> {
   return manualDownloadRequired
 }
 
-async function setupActiveJar (minecraftVersion: string) {
+async function setupActiveJar (minecraftVersion: string): Promise<void> {
   const builds = await getBuilds(minecraftVersion)
   if (!builds) throw new Error(`No builds available for ${minecraftVersion}`)
 
@@ -130,15 +160,94 @@ async function setupActiveJar (minecraftVersion: string) {
     latestBuild.downloads.application.sha256
   )
 
-  if (!haveValidBuild) await downloadBuildJar(minecraftVersion, latestBuild.build)
+  console.log('Latest paper version:', latestBuild.downloads.application.name)
+  console.log('Needs downloading:', !haveValidBuild)
+
+  if (!haveValidBuild) {
+    await downloadBuildJar(minecraftVersion, latestBuild.build)
+    console.log('Downloaded latest paper version')
+  }
 
   if (fs.existsSync('./server/paper.jar')) fs.rmSync('./server/paper.jar')
 
-  fs.copyFileSync(`./bin/papermc/${latestBuild.downloads.application.name}`, './server/paper.jar')
+  console.log('Copying', latestBuild.downloads.application.name, 'to ./server/paper.jar')
+  return new Promise((resolve, reject) => {
+    fs.copyFile(`./bin/papermc/${latestBuild.downloads.application.name}`, './server/paper.jar', (error) => {
+      console.debug('copy errors:', {error})
+      if (error) reject(error)
+      resolve()
+    })
+  })
+}
+
+async function checkForOpenJRE(): Promise<string | void> {
+  if (!fs.existsSync('./bin/openjre')) fs.mkdirSync('./bin/openjre')
+
+  const _arch = process.arch
+  const _platform = process.platform
+
+  let arch: adoptArch
+  let platform: adoptOS
+
+  switch (_arch) {
+    case "arm64":
+    case "ia32":
+    case "mips":
+    case "mipsel":
+    case "ppc":
+    case "s390":
+      console.error(_arch, 'architecture is not supported by AdoptOpenJDK')
+      return
+    default:
+      arch = _arch
+      break
+  }
+
+  switch (_platform) {
+    case "win32":
+      platform = "windows"
+      break
+    case "android":
+    case "darwin":
+    case "freebsd":
+    case "haiku":
+    case "openbsd":
+    case "sunos":
+    case "cygwin":
+    case "netbsd":
+      console.error(_platform, 'platform is not supported by AdoptOpenJDK')
+      return
+    default:
+      platform = _platform
+      break
+  }
+
+  const jreAsset = await getAssets(arch, platform)
+
+  if (!jreAsset) {
+    console.error('Unable to find available JRE for', platform, arch)
+    return
+  }
+
+  const jrePath = jreAsset.release_name + '-jre'
+
+  if (!fs.existsSync('./bin/openjre/' + jrePath)) {
+    const packagePath = './bin/openjre/' + jreAsset.binary.package.name
+    await downloadJRE(jreAsset.binary.package.link, jreAsset.binary.package.name)
+    await decompress('./bin/openjre/' + jreAsset.binary.package.name, './bin/openjre/')
+    fs.rmSync(packagePath)
+  }
+
+  return jrePath
+
 }
 
 // Setup up paper jar we gonna be using
-setupActiveJar(minecraftVersion).then(async () => {
+setupActiveJar(config.minecraft_version).then(async () => {
+  const javaPath = await checkForOpenJRE()
+
+  if (!javaPath) throw new Error('Unable to find java exec, wah wah')
+
   fs.readdir('./bin/plugins', (err, files) => {
     for (const file of files) {
       fs.rmSync(`./bin/plugins/` + file)
@@ -146,39 +255,55 @@ setupActiveJar(minecraftVersion).then(async () => {
   })
 
   const essentialsBuild = await checkForLatestEssentialPlugins()
-  const enabledPlugins = (process.env.ESSENTIALSX_PLUGINS || '').split(',')
 
   essentialsBuild.assets.forEach(asset => {
-    if (enabledPlugins.some(enabledAsset => {
-      return asset.name.startsWith(enabledAsset + '-')
-    })) downloadAssetJar(essentialsBuild.tag_name, asset.name)
+    const isEssentialsAssetEnabled = config.plugins.essentials_x.some(enabledAsset => asset.name.startsWith(enabledAsset + '-'))
+    if (isEssentialsAssetEnabled) {
+      console.log('Downloading', asset.name)
+      downloadAssetJar(essentialsBuild.tag_name, asset.name)
+    }
   })
 
   const { latestGeyser, latestFloodgate } = await checkForLatestGeyser()
 
+  console.log('Downloading Geyser', latestGeyser.version, latestGeyser.build)
   await downloadGeyser(latestGeyser.version, latestGeyser.build)
+
+  console.log('Downloading Floodgate', latestFloodgate.version, latestFloodgate.build)
   await downloadFloodgate(latestFloodgate.version, latestFloodgate.build)
 
   const waitBeforeStart = await checkForOtherPlugins()
   if (!fs.existsSync('./server/plugins')) fs.mkdirSync('./server/plugins')
+
+  fs.readdir('./server/plugins/', (err, files) => {
+    for (const file of files) {
+      if (file.endsWith('.jar')) fs.rmSync('./server/plugins/' + file)
+    }
+  })
+
   fs.readdir('./bin/plugins', (err, files) => {
     for (const file of files) {
       fs.copyFileSync('./bin/plugins/' + file, './server/plugins/' + file, fs.constants.COPYFILE_FICLONE)
     }
   })
 
-  fs.writeFileSync('./server/eula.txt','#suck it\n#microsoft\neula=false') // Is this legal??
+  fs.writeFileSync('./server/eula.txt','#suck it\n#microsoft\neula=true') // Is this legal??
 
-  await new Promise((r) => { setTimeout(r, 1000) })
+  if (waitBeforeStart) {
+    console.log('[PAUSED] press ENTER to continue...')
+    await waitForKey()
+  }
 
-  fs.writeFileSync('./server/eula.txt', '#suck it\n#microsoft\neula=true') // Probably not lol
+  console.log(textSync('Starting Server', "Big Money-se").rainbow)
 
-  console.log('[PAUSED] press ENTER to continue...')
-  await waitForKey()
+  let javaExec = '../bin/openjre/' + javaPath + '/bin/java'
+  switch (process.platform) {
+    case "win32":
+      javaExec = javaExec + '.exe'
+      break
+  }
 
-  console.log('Starting Server')
-
-  const child = spawn('java', [
+  const child = spawn(javaExec, [
     '-Xms5000M',
     '-Xmx5000M',
     '-jar',
@@ -193,8 +318,39 @@ setupActiveJar(minecraftVersion).then(async () => {
     ]
   })
 
+  const timestampRegex = /^\[\d{2}:\d{2}:\d{2} \D{4}]/is // [00:00:00 INFO]
+  const titleRegex = /\[\D+?]/gi // [Essentials] [Vault] [Harbor] etc
+
   child.stdout.on('data', data => {
-    console.log(data.toString())
+    let messageString = data.toString()
+    if (messageString.endsWith('\n')) messageString = messageString.substring(0, messageString.length - 1)
+
+    messageString.split('\n').forEach((message: string) => {
+      const isLogMessage = timestampRegex.test(message)
+
+      const msgObj = {
+        type: isLogMessage ? message.substring(10, 14) : 'SYSTEM',
+        message: isLogMessage ? message.substring(17) : message,
+        raw: message
+      }
+
+      if (msgObj.type === 'SYSTEM') msgObj.message = msgObj.message.bgYellow.black
+      else if (msgObj.type === 'WARN') msgObj.message = msgObj.message.yellow
+      else if (msgObj.type === 'ERRO') msgObj.message = msgObj.message.red
+
+      else if (msgObj.message.startsWith('Environment: ')) msgObj.message = msgObj.message.bgMagenta
+      else if (msgObj.message.endsWith('For help, type "help"')) {
+        console.log(textSync('Server Started!', "Big Money-ne").yellow)
+      }
+
+      if (titleRegex.test(msgObj.message)) {
+        msgObj.message = msgObj.message.replaceAll(titleRegex, (match) => {
+          return match.bgCyan.black
+        })
+      }
+
+      console.log(msgObj.message)
+    })
   })
 
   child.on('exit', (code, signal) => {
